Games controller rspec tests

+++ app/controllers/games_controller.rb
@@ -1,7 +1,7 @@
 class GamesController < ApplicationController
   include GamesHelper
 
-  before_action :authenticate_user!, only: [:new, :create]
+  before_action :authenticate_user!, only: [:new, :create, :edit, :update, :destroy]
 
   def index
     @games = Game.all
@@ -22,26 +22,38 @@ class GamesController < ApplicationController
   end
 
   def show
+    return render_not_found if current_game.blank?
   end
 
   def edit
+    return render_not_found(:forbidden) if current_user.id != current_game.user_id
   end
 
+  # to join game as another user, or update game name if user created game
+  # rubocop:disable Metrics/AbcSize
   def update
-    if user_signed_in? && current_user.id != current_game.user_id
+    if current_user.id != current_game.user_id && game_is_open
       current_game.update_attributes(game_params)
       flash.notice = 'You have successfully joined the game!'
       redirect_to game_path(current_game)
-    elsif user_signed_in?
-      flash.alert = 'You are the host of this game!'
-      redirect_to root_path
+    elsif current_user.id == current_game.user_id
+      if params[:on_edit_page] == 'yes'
+        current_game.update_attributes(game_params)
+        flash.notice = 'Game name changed!'
+        redirect_to game_path(current_game)
+      else
+        flash.alert = 'You are the host of this game!'
+        redirect_to root_path
+      end
     else
-      flash.alert = 'You must be signed in to join games!'
-      redirect_to root_path
+      return render_not_found(:forbidden)
     end
   end
+  # rubocop:enable Metrics/AbcSize
 
   def destroy
+    return render_not_found(:forbidden) if current_user.id != current_game.user_id
+    return render_not_found(:forbidden) unless game_is_open
     current_game.destroy
     redirect_to root_path
   end
@@ -51,14 +63,18 @@ class GamesController < ApplicationController
   #   redirect_to game_path(current_game)
   # end
 
+  private
+
   def current_game
     @current_game = Game.find(params[:id])
   end
   helper_method :current_game
 
-  private
-
   def game_params
     params.require(:game).permit(:name, :player_black_id, :player_white_id, :status, :user_id, :turn)
+  end
+
+  def game_is_open
+    current_game.player_white_id.nil? || current_game.player_black_id.nil?
   end
 end
